working authentication

+++ storage-ms/app/api/item_api.py
@@ -2,8 +2,12 @@
 # Date created: 5.12.2024
 
 # REST FastAPI dependencies.
-from fastapi import APIRouter, HTTPException
+from fastapi import APIRouter, HTTPException, Depends
 from fastapi.responses import PlainTextResponse
+
+# OAuth2 authentication dependencies.
+from auth.token_bearer import JWTBearer
+from auth.token_utils import validate_token_with_username
 
 # Internal dependencies.
 from ..schemas import item_schemas as schema
@@ -11,13 +15,15 @@ from ..models.item import Item
 from ..services import item_utils as utils
 from ..helpers.error import ErrorResponse as Err
 
+token_bearer = JWTBearer()
+
 router = APIRouter(
     prefix="/users",
     tags=["users"]
 )
 
 @router.post("/{username}/{storage_name}/create-item", status_code=200, response_class=PlainTextResponse)
-async def create_item(username: str, storage_name: str, item_schema: schema.ItemCreate):
+async def create_item(username: str, storage_name: str, item_schema: schema.ItemCreate, token : str = Depends(token_bearer)):
     """
     This endpoint allows the creation of a new item within a specific storage for a user.
 
@@ -25,6 +31,8 @@ async def create_item(username: str, storage_name: str, item_schema: schema.Item
         username (str): The username of the user.
         storage_name (str): The name of the storage where the item will be created.
         item_schema (ItemCreate): The item details to be created.
+        token (str): Access token generated at login time.
+
 
     Raises:
         HTTPException: If an error occurs during item creation.
@@ -33,6 +41,9 @@ async def create_item(username: str, storage_name: str, item_schema: schema.Item
         PlainTextResponse: The item code if the item is created successfully.
     """
 
+    if not validate_token_with_username(username, token):
+        raise HTTPException(status_code=401, detail="Token username missmatch.")
+
     result = await utils.create_item(username, storage_name, item_schema)
     if isinstance(result, Err):
         raise HTTPException(status_code=result.code, detail=result.message)
@@ -40,7 +51,7 @@ async def create_item(username: str, storage_name: str, item_schema: schema.Item
     return result
 
 @router.get("/{username}/{storage_name}/{item_code}", response_model=Item)
-async def get_item(username: str, storage_name: str, item_code: str):
+async def get_item(username: str, storage_name: str, item_code: str, token : str = Depends(token_bearer)):
     """
     This endpoint fetches an item by its unique code from a user's specified storage.
 
@@ -48,6 +59,7 @@ async def get_item(username: str, storage_name: str, item_code: str):
         username (str): The username of the user.
         storage_name (str): The name of the storage containing the item.
         item_code (str): The unique code of the item.
+        token (str): Access token generated at login time.
 
     Raises:
         HTTPException: If an error occurs during item retrieval.
@@ -56,6 +68,9 @@ async def get_item(username: str, storage_name: str, item_code: str):
         Item: The retrieved item details.
     """
 
+    if not validate_token_with_username(username, token):
+        raise HTTPException(status_code=401, detail="Token username missmatch.")
+
     result = await utils.get_item(username, storage_name, item_code)
     if isinstance(result, Err):
         raise HTTPException(status_code=result.code, detail=result.message)
@@ -63,7 +78,7 @@ async def get_item(username: str, storage_name: str, item_code: str):
     return result
 
 @router.delete("/{username}/{storage_name}/{item_code}", status_code=200, response_class=PlainTextResponse)
-async def delete_item(username: str, storage_name: str, item_code: str):
+async def delete_item(username: str, storage_name: str, item_code: str, token : str = Depends(token_bearer)):
     """
     Delete an item from a user's storage.
 
@@ -73,6 +88,7 @@ async def delete_item(username: str, storage_name: str, item_code: str):
         username (str): The identifier of the user.
         storage_name (str): The name of the storage containing the item.
         item_code (str): The unique code of the item to delete.
+        token (str): Access token generated at login time.
 
     Raises:
         HTTPException: If an error occurs during item deletion.
@@ -81,6 +97,9 @@ async def delete_item(username: str, storage_name: str, item_code: str):
         PlainTextResponse: The item code if the item is deleted successfully.
     """
 
+    if not validate_token_with_username(username, token):
+        raise HTTPException(status_code=401, detail="Token username missmatch.")
+
     result = await utils.delete_item(username, storage_name, item_code)
     if isinstance(result, Err):
         raise HTTPException(status_code=result.code, detail=result.message)
@@ -88,7 +107,7 @@ async def delete_item(username: str, storage_name: str, item_code: str):
     return result
 
 @router.put("/{username}/{storage_name}/{item_code}", status_code=200, response_class=PlainTextResponse)
-async def update_item(username: str, storage_name: str,  item_code: str, item : schema.ItemUpdate):
+async def update_item(username: str, storage_name: str,  item_code: str, item : schema.ItemUpdate, token : str = Depends(token_bearer)):
     """
     This endpoint allows updating the details of an existing item in a specified storage.
 
@@ -97,6 +116,7 @@ async def update_item(username: str, storage_name: str,  item_code: str, item : 
         storage_name (str): The name of the storage containing the item.
         item_code (str): The unique code of the item to update.
         item (ItemUpdate): The updated item details.
+        token (str): Access token generated at login time.
 
     Raises:
         HTTPException: If an error occurs during item update.
@@ -104,6 +124,9 @@ async def update_item(username: str, storage_name: str,  item_code: str, item : 
     Returns:
         PlainTextResponse: The item code if the item is updated successfully.
     """
+
+    if not validate_token_with_username(username, token):
+        raise HTTPException(status_code=401, detail="Token username missmatch.")
 
     result = await utils.update_item(username, storage_name, item_code, item)
     if isinstance(result, Err):
