+++ storage-ms/app/api/users_api.py
@@ -6,18 +6,14 @@ from fastapi import APIRouter, HTTPException, Depends
 from fastapi.responses import PlainTextResponse
 
 # OAuth2 authentication dependencies.
-from typing import Annotated
-from fastapi.security import (
-    OAuth2PasswordBearer)
 from auth.token_bearer import JWTBearer
-from auth.token_parser import validate_token_with_username
+from auth.token_utils import validate_token_with_username
 
 # Internal dependencies.
 from ..schemas import user_schemas as schema
 from ..services import user_utils as utils
 from ..models.user import User
 from ..helpers.error import ErrorResponse as Err
-from auth.token_utils import validate_token_with_username
 
 token_bearer = JWTBearer()
 
@@ -41,6 +37,7 @@ async def create_user(user_schema : schema.UserCreate, token : str = Depends(tok
     Returns:
         PlainTextResponse: The username if the user is created successfully.
     """
+
     if not validate_token_with_username(username, token):
         raise HTTPException(status_code=401, detail="Token username missmatch.")
 
@@ -65,6 +62,7 @@ async def get_user(username: str, token : str = Depends(token_bearer)):
     Returns:
         User: The retrieved user details.
     """
+
     if not validate_token_with_username(username, token):
         raise HTTPException(status_code=401, detail="Token username missmatch.")
 
@@ -75,12 +73,13 @@ async def get_user(username: str, token : str = Depends(token_bearer)):
     return result
 
 @router.delete("/{username}", status_code=200, response_class=PlainTextResponse)
-async def delete_user(username: str):
+async def delete_user(username: str, token : str = Depends(token_bearer)):
     """
     This endpoint removes a user from the system by their username.
 
     Args:
         username (str): The username of the user to delete.
+        token (str): Access token generated at login time.
 
     Raises:
         HTTPException: If an error occurs during user deletion.
@@ -89,6 +88,9 @@ async def delete_user(username: str):
         PlainTextResponse: The username if the user is deleted successfully.
     """
 
+    if not validate_token_with_username(username, token):
+        raise HTTPException(status_code=401, detail="Token username missmatch.")
+
     result = await utils.delete_user(username)
     if isinstance(result, Err):
         raise HTTPException(status_code=result.code, detail=result.message)
@@ -96,13 +98,14 @@ async def delete_user(username: str):
     return result
 
 @router.put("/{username}/update-name", status_code=200, response_class=PlainTextResponse)
-async def update_display_name(username: str, new_name : str):
+async def update_display_name(username: str, new_name : str, token : str = Depends(token_bearer)):
     """
     This endpoint allows updating the display name of a user.
 
     Args:
         username (str): The username of the user to update.
         new_name (str): The new display name for the user.
+        token (str): Access token generated at login time.
 
     Raises:
         HTTPException: If an error occurs during display name update.
@@ -111,6 +114,9 @@ async def update_display_name(username: str, new_name : str):
         PlainTextResponse: The username if the users display name is updated successfully.
     """
 
+    if not validate_token_with_username(username, token):
+        raise HTTPException(status_code=401, detail="Token username missmatch.")
+
     result = await utils.update_display_name(username, new_name)
     if isinstance(result, Err):
         raise HTTPException(status_code=result.code, detail=result.message)
@@ -118,12 +124,13 @@ async def update_display_name(username: str, new_name : str):
     return result
 
 @router.put("/{username}/empty-storages", status_code=200, response_class=PlainTextResponse)
-async def empty_storages(username: str):
+async def empty_storages(username: str, token : str = Depends(token_bearer)):
     """
     This endpoint clears all storages in a user's account.
 
     Args:
         username (str): The username of the user whose storages will be emptied.
+        token (str): Access token generated at login time.
 
     Raises:
         HTTPException: If an error occurs during storages clearing.
@@ -131,6 +138,9 @@ async def empty_storages(username: str):
     Returns:
         PlainTextResponse: The username if the user storages are emptied successfully.
     """
+
+    if not validate_token_with_username(username, token):
+        raise HTTPException(status_code=401, detail="Token username missmatch.")
 
     result = await utils.empty_storages(username)
     if isinstance(result, Err):
